Fix block light properties
+ Shapes now hold default values for neighborBrightness, translucent and
lightOpacity. This makes it so these values don't have to be specified
when defining a slab for example. The value we get IN from JSON is
checked against it's shape value so it's not written to it either. If
one of these properties ARE defined in a JSON block, we use that value
instead of the default, just like before.
+ Textures are no longer forcibly moved if it already exists
+ Added some debugger code. When the player is in creative and right
clicks a block with an empty hand, the user is given information through
the chat about that block. Image http://i.imgur.com/WeFwaV2.png

+++ src/main/java/com/kimfy/notenoughblocks/common/file/resourcepack/ResourcePack.java
@@ -111,6 +111,7 @@ public class ResourcePack
         FileUtilities.extractZip(file, Constants.PATH_MOD_CONFIG_RESOURCE_PACKS + name + "_extracted");
     }
 
+    // TODO: Implement hasAlreadyBeenMoved and cache the results
     public void move()
     {
         File temp = new File(Constants.PATH_MOD_CONFIG_RESOURCE_PACKS + name + "_extracted/assets/minecraft/textures/blocks/");
@@ -128,7 +129,8 @@ public class ResourcePack
                         String textureName = name + "_" + file.getName();
                         FileManager.textures.add(textureName.replace(".png", ""));
                         File destinationFile = new File(destination + textureName/*name + "_" + file.getName()*/);
-                        Files.move(file, destinationFile);
+                        if (!destinationFile.exists())
+                            Files.move(file, destinationFile);
                     }
                     catch (IOException e)
                     {
